Disable Disk-based Shard Allocation
Disk space that is allocated to Elasticsearch is controlled by securityonion.conf and curator.  If Disk-based Shard Allocation is enabled, it leads to disk watermark errors and indices being locked as readonly.  Since Security Onion configures each Elasticsearch instance as a single node cluster, the index can never be moved to another node.  This results in data being lost and not ingested into Elasticsearch.